Quick fix de debug de nvx formats de données

+++ general_process/SourceProcess.py
@@ -77,8 +77,10 @@ class SourceProcess:
         self.file_name = [f"{self.metadata[self.key]['code']}_{i}" for i in range(len(self.url))]
         os.makedirs(f"sources/{self.source}", exist_ok=True)
         for i in range(len(self.url)):
-            print(self.url[i])
-            wget.download(self.url[i], f"sources/{self.source}/{self.file_name[i]}.{self.format}")
+            try:
+                wget.download(self.url[i], f"sources/{self.source}/{self.file_name[i]}.{self.format}")
+            except:
+                print("Problème de téléchargement du fichier ", self.url[i])
         logging.info(f"Téléchargement : {len(self.url)} fichier(s) OK")
 
     def convert(self):
@@ -93,10 +95,14 @@ class SourceProcess:
                 count += 1
         for i in range(count):
             file_path = f"sources/{self.source}/{self.file_name[i]}.{self.format}"
-            with open(file_path) as file:
-                chaine = re.sub("\'", " ", file.read())
-            with open(file_path, "w") as file:
-                file.write(chaine)
+            file_exist = os.path.exists(file_path)
+            if file_exist:
+                with open(file_path) as file:
+                    chaine = re.sub("\'", " ", file.read())
+                with open(file_path, "w") as file:
+                    file.write(chaine)
+            else:
+                print(f" Fichier {file_path} n existe pas.")
         if count != len(self.url):
             logging.warning("Nombre de fichiers en local inégal au nombre d'url trouvé")
         logging.info(f"Début de convert: mise au format DataFrame de {self.source}")
@@ -128,10 +134,13 @@ class SourceProcess:
         elif self.format == 'json':
             li = []
             for i in range(count):
-                with open(
-                        f"sources/{self.source}/{self.file_name[i]}.{self.format}") as json_file:
-                    dico = json.load(json_file)
-                li.append(pd.DataFrame.from_dict(dico['marches']))
+                try:
+                    with open(
+                            f"sources/{self.source}/{self.file_name[i]}.{self.format}") as json_file:
+                        dico = json.load(json_file)
+                        li.append(pd.DataFrame.from_dict(dico['marches']))
+                except:
+                    print(f" Le fichier {self.source} est introuvable")
             df = pd.concat(li)
             df = df.reset_index(drop=True)
             self.df = df
